OD-10838 Implement 'addTheme' api
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@49323 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ editor-api/src/main/groovy/ish/oncourse/willow/editor/rest/WebNodeTypeToTheme.groovy
@@ -25,7 +25,7 @@ class WebNodeTypeToTheme {
         return new Theme().with { theme ->
             theme.title = webNodeType.name
             theme.id = webNodeType.id.doubleValue()
-            theme.layoutId = webNodeType.webSiteLayout.id
+            theme.layoutId = webNodeType.webSiteLayout.id.doubleValue()
             theme.schema = new ThemeSchema().with { schema ->
                 schema.top = getBlocksBy(header)
                 schema.left = getBlocksBy(left)
